update all features

+++ app/Http/Controllers/LowonganController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Lowongan;
 use App\Models\Perusahaan;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class LowonganController extends Controller
@@ -15,15 +16,19 @@ class LowonganController extends Controller
      */
     public function index()
     {
-        $title_alert = 'Hapus Data Lowongan';
-        $text_alert = "Apakah anda yakin menghapus data ini?";
-        confirmDelete($title_alert, $text_alert);
+        Gate::authorize('viewAny', Lowongan::class);
+
+        $title = 'Delete User!';
+        $text = "Are you sure you want to delete?";
+        confirmDelete($title, $text);
+
+        $lowongan = Lowongan::latest()->get();
 
         return view(
             'admin.lowongan.index',
             [
                 'title' => 'Lowongan Pekerjaan',
-                'data' => Lowongan::all(),
+                'data' => $lowongan,
             ]
         );
     }
@@ -33,7 +38,9 @@ class LowonganController extends Controller
      */
     public function create()
     {
-        $perusahaan = Perusahaan::all();
+        Gate::authorize('create', Lowongan::class);
+
+        $perusahaan = Perusahaan::first();
         return view(
             'admin.lowongan.create',
             [
@@ -48,6 +55,8 @@ class LowonganController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Lowongan::class);
+
         $request->validate(
             [
                 'perusahaan_id' => 'required',
@@ -55,7 +64,9 @@ class LowonganController extends Controller
                 'slug' => 'required|unique:lowongan,slug',
                 'kategori' => 'required',
                 'tipe' => 'required',
+                'informasi' => 'required',
                 'deskripsi' => 'required',
+                'batas_akhir' => 'required',
                 'status' => 'required',
                 'gambar' => 'required',
                 'gambar.*' => 'image|mimes:jpeg,png,jpg,svg,webp',
@@ -68,21 +79,25 @@ class LowonganController extends Controller
                 'slug.unique' => 'Slug sudah terdaftar, mohon pilih slug yang lain',
                 'kategori.required' => 'Inputan kategori harus diisi',
                 'tipe.required' => 'Inputan tipe harus diisi',
+                'informasi.required' => 'Inputan informasi lowongan harus diisi',
                 'deskripsi.required' => 'Inputan deskripsi harus diisi',
+                'batas_akhir.required' => 'Inputan batas akhir harus diisi',
                 'status.required' => 'Inputan status harus diisi',
-                'gambar.required' => 'Inputan gambar harus diisi',
-                'gambar.image' => 'File gambar harus diisi dengan file jpeg, png, jpg, svg, webp',
+                'gambar.required' => 'Inputan poster harus diisi',
+                'gambar.image' => 'File poster harus diisi dengan file jpeg, png, jpg, svg, webp',
             ]
         );
 
         $input = $request->all();
 
-        if ($request->hasFile("gambar")) {
-            $image = $request->file("gambar");
-            $destinationPath = "media/";
-            $profileImage = date("YmdHis") . "." . $image->getClientOriginalExtension();
-            $image->move($destinationPath, $profileImage);
-            $input["gambar"] = "$profileImage";
+        if ($request->hasFile('gambar')) {
+            $image = $request->file('gambar');
+
+            $path = $image->storeAs('public/lowongan', date('YmdHis') . '.' . $image->getClientOriginalExtension());
+
+            $link = Storage::url($path);
+
+            $input['gambar'] = $link;
         }
 
         Lowongan::create($input);
@@ -96,7 +111,15 @@ class LowonganController extends Controller
      */
     public function show(Lowongan $lowongan)
     {
-        //
+        Gate::authorize('view', $lowongan);
+
+        return view(
+            'admin.lowongan.detail',
+            [
+                'title' => 'Lowongan Pekerjaan',
+                'data' => $lowongan,
+            ]
+        );
     }
 
     /**
@@ -104,6 +127,8 @@ class LowonganController extends Controller
      */
     public function edit(Lowongan $lowongan)
     {
+        Gate::authorize('update', $lowongan);
+
         $perusahaan = Perusahaan::where('id', $lowongan->perusahaan_id)->first();
 
         return view(
@@ -121,11 +146,14 @@ class LowonganController extends Controller
      */
     public function update(Request $request, Lowongan $lowongan)
     {
+        Gate::authorize('update', $lowongan);
+
         $request->validate(
             [
                 'perusahaan_id' => 'required',
                 'judul' => 'required',
                 'slug' => 'required',
+                'informasi' => 'required',
                 'kategori' => 'required',
                 'tipe' => 'required',
                 'deskripsi' => 'required',
@@ -136,26 +164,28 @@ class LowonganController extends Controller
                 'perusahaan_id.required' => 'Inputan perusahaan harus diisi',
                 'judul.required' => 'Inputan judul harus diisi',
                 'slug.required' => 'Inputan slug harus diisi',
+                'informasi.required' => 'Inputan informasi lowongan harus diisi',
                 'kategori.required' => 'Inputan kategori harus diisi',
                 'tipe.required' => 'Inputan tipe harus diisi',
                 'deskripsi.required' => 'Inputan deskripsi harus diisi',
                 'status.required' => 'Inputan status harus diisi',
-                'gambar.image' => 'File gambar harus diisi dengan file jpeg, png, jpg, svg, webp',
+                'gambar.image' => 'File poster harus diisi dengan file jpeg, png, jpg, svg, webp',
             ]
         );
 
         $input = $request->all();
 
-        if ($request->hasFile("gambar")) {
-            File::delete('media/' . $lowongan->gambar);
+        if ($request->hasFile('gambar')) {
+            if ($lowongan->gambar) {
+                Storage::delete('public/profile/' . basename($lowongan->gambar));
+            }
+            $image = $request->file('gambar');
 
-            $image = $request->file("gambar");
-            $destinationPath = "media/";
-            $profileImage = date("YmdHis") . "." . $image->getClientOriginalExtension();
-            $image->move($destinationPath, $profileImage);
-            $input["gambar"] = "$profileImage";
-        } else {
-            unset($input["gambar"]);
+            $path = $image->storeAs('public/lowongan', date('YmdHis') . '.' . $image->getClientOriginalExtension());
+
+            $link = Storage::url($path);
+
+            $input['gambar'] = $link;
         }
 
         $lowongan->update($input);
@@ -169,7 +199,9 @@ class LowonganController extends Controller
      */
     public function destroy(Lowongan $lowongan)
     {
-        File::delete('media/' . $lowongan->gambar);
+        Gate::authorize('delete', $lowongan);
+
+        Storage::delete('public/profile/' . basename($lowongan->gambar));
 
         $lowongan->delete();
